Turn sendfile off by default in nginx
The default for sendfile is on, which appears to not jive with something about the VM for some things, causing weird character encoding issues in Javascript (magic syntax errors abound) and character encoding issues in CSS (makes it seem like files are cached forever) that are just no fun to deal with.

Setting to 'off' for now until I learn more about what sendfile does and how important it might be to have in a local environment.