Remove current implementation of xdebug support
The xdebug_on and xdebug_off commands were not working properly as there was no config file available for PHP to be used with xdebug_off. Disabling xdebug entirely now to focus on some other changes. We can bring this back in a bit later.

Can I tag people in GitHub? @carldanley