Move to prepackaged 10up-quantal-0.1 as the base box
Our 10up-quantal-0.1 base box provides included provisioning for nginx, mysql, memcached and php5-fpm as well as 4.2.6 Virtual Box Guest Additions. By repackaging this common configuration, we save some time in the provisioning process and require less from the outside world during the build.

I'd like to leave PECL memcache out of the package, but I have a feeling I'm about to recreate a package so that no internet access is needed during this provisioning.

Old provisioning information - or what the box is packaged with - is included in a new provision-details.txt file for reference.