Introduce watchr as a method to execute scripts upon file modification
It'd be great if we didn't have to destroy a box and bring it back up just to add new web projects. To do this, we want to monitor some files and look for any changes so that they can immediately be brought into the environment.

watchr allows us to do exactly this.

A .bash_aliases file is copied over, as the default Vagrant .bashrc automatically includes that if it exists on boot. This avoids our having to maintain an entire bashrc file

The watchr.script file will probably expand to look for changes on a few things. Right now we monitor the initial DB file and then run the import script if anything changes. We'll definitely want to get a bit smarter on how that's handled, but this is a start.