Add a default web directory to serve at the IP address
The best way to make sure this is working without any other sites configured is to have some sort of default configuration that responds at the IP address.

The server config will need to change if you intend to use something different than 192.168.50.4, though we can probably work out a templating system in the future.