Add piped comments after every file copy to avoid messy line endings
This is cheap, but I think effective. In certain cases on Windows machines, it's possible that the line endings in the provision.sh get ugly. When the copy commands are run to move different c

By adding a piped message to the end, we make sure that the filename is not what ends the statement. This should avoid any kind of weird filenames being copied into our config directories.