Add an mu-plugin that can be used with WordPress to alter Cron requests
Something about either our Ubuntu setup or the fact that things are happening in a virtual machine is causing problems with the default 0.01 timeout used by WordPress for cron requests. Cron appears to still work fine, but a notice comes up regarding the fopen error.

This mu-plugins folder can be dropped in to any WordPress wp-content directory and will automatically alter the cron timeout to avoid notices