Remove stored procedure due to wth
I didn't think this through the right way. We don't need a stored procedure, we just need a command that doesn't error out.

By using GRANT PRIVILEGES without the CREATE USER command, we ensure that privileges are assigned whether the user already exists or not. If it doesn't, it will automatically be created for us with no error.