Only run apt-get update on first boot
apt-get update is the part of the `vagrant up` process after a `vagrant halt` that takes the longest because it will go out and update all of the source lists. Because we are not installing anything extra, we can skip this if it has run before.

On first boot, we create an empty `apt_update_run` file in our /srv/config directory. Subsequent boots will see this file and then skip the update command