Clean up apt-get installs in provisioning script
The provision script was so cleverly split up into many `apt-get install` statements, which was great but caused a little bit of a delay in processing each time so that it could gather the package data.

This commit combines all of those packages into a bash array that we pass to one `apt-get install` command so that the entire shebang can be completed at once.

In additon, some whitespace stuff has been cleaned up in the output because pretty output is better than ugly output.

Closes #20