Support database names with hyphens in import-sql script
The bash script that helps us import backup SQL files into existing databases that have been created with init-custom.sql was choking on databases with hypens in the name. This performs a little magic by assigning the SQL statement to a variable first before going about the commands as normal.

Fixes #19