Stop watching init-custom.sql for changes
It doesn't make sense, especially with our new data structure, to reimport all DBs whenever the init-custom.sql file has been modified. We're planning for a smarter way of importing new projects in the future, but for now we'll have to live with mysql commands to create databases after a vagrant up - or... just halt the box and bring it back up again.