Go through full installation of WordPress default site
Rather than leave the site hanging for details, configure it and install it. This does require moving this section of code down below the service restart area so that mysql is available for wp-cli to use.

Next, we'll see what we can do about getting this data from a project config file