Stop the apt-get update flag for 0.3
The check to see if apt-get update had run in the past is a good idea, but fails in practice if you destroy the vagrant box and then run vagrant up again from time to time due to the package source files expiring.

We'll solve this in a different way in the future, but for now it's too shaky to include.