Skip all package installation in provisioning after initial boot
Creates a /home/vagrant/initial_provision_run file to track once provisioning has been run on a box for the first time. If this file exists, all apt and composer packages are skipped. Only file linking and WordPress management remains.

If the box is ever destroyed, this flag will be deleted and provisioning will run on the next boot.

See #40