Use the powers of Ruby to provide for a custom Vagrantfile
As the initial Vagrantfile loads, we check for the existence of a Customfile that can contain additional Vagrant configuration parameters. If this file does not exist on your local system, nothing is loaded.

May be useful especially for mapping other drives on the system to setup advanced configurations.

Will need to test whether it will overwrite previously set configuration options in the Vagrantfile