Append a custom sources list to the default provided for Apt
Some of the packages included by default with the sources list in Ubuntu 12.04 are way out of date, or boring at the least. This commit allows custom sources to be appended to the list so that we can better control the versions that are installed.

A specific change from this commit is the use of SVN 1.7.9, which should help to resolve file permissions issues on some Windows machines.

Closes #58