Up permissions on files in srv/www to 774 so that wp-cli can execute
This is a blanket fix and should probably be thought of more as we go. It doesn't really cause a security issue because this is all happening inside a VM on your local machine, but it's still worth thinking about.

One option is to not explicitly set dmode and fmode when mapping the directories and then handling that part manually through creative uses of chmod.