Provide a wp-config-sample.php that enables WP_DEBUG during a wp-cli controlled installation
wp-cli uses the wp-config-sample.php file inside a WordPress directory to setup WordPress and leaves the WP_DEBUG value as it found it. By providing our own wp-config-sample.php, we can control this a bit more.

This setups up a default mu-plugins directory as well that we should then copy to new installations.