Refactor handling of xdebug config
Because the xdebug.so location can change whenever an updated version of PHP comes around, we want to try to avoid specifying that location directly in the php.ini file. By default, when installed with `apt-get`, xdebug has its own ini file in `/etc/php5/fpm/conf.d/` that loads the extension as required.

We can then come in and use an additional .ini file in that directory that accounts for the remainder of the xdebug configuration.

Committing fairly blindly for now but will test and tweak near term.