Remove additional xdebug configs in favor of php5enmod and php5dismod
Our xdebug_off and xdebug_on commands in the bash_aliases file depended on an xdebug configuration that wanted xdebug.so to be in one place. Unfortunately we were at the mercy of PHP as to where that place was.

By creating a separate xdebug.ini file that does not load the extension information, we can turn to php5enmod and php5dismod to enable and disable xdebug instead while always having the xdebug.ini with remaining configuration available to us.

This does bring into question the usefulness of xdebug_off and xdebug_on, but I'll create a new issue for that.