Change nginx.conf to only include .conf files
I'm doing this in a terrible way at the moment because I'm not being thorough and fixing the things this breaks - but I'd like to change this to include only *.conf files in the Nginx directories.  This is specifically to avoid /etc/nginx/custom-sites/local-nginx-example.conf-sample from being loaded in Nginx.  However, it also breaks the "default" files - which I'd recommend changing to default.conf instead.  The benefit of this is you can quickly backup a conf file by copying it to conf.bak and it won't get pulled in to Nginx.  I can be more thorough about this later.