Handle package status in provision.sh for repeat provisions
This treats provision.sh more like the provisioning script it should be and checks for the status of each package before adding it to the array of apt packages to be installed. If the array is not empty, `apt-get update` runs and then the `apt-get install` immediately after.

Sets us up to remove the initial provision flag, which was an ugly hack. Better yet, if a package fails to install the first time round, a quick `vagrant provision` should take care of it.