Provide our own default apt sources.list to append our custom sources to
The previous append method was adding our custom sources to the default sources.list every time that `vagrant provision` was run, causing many duplicate entries. Rather than deal with the logic of checking for existing lines in that file, we'll just take control of the entire thing by providing the default sources from Ubuntu 12.04 ourselves. There may be a better long term fix for this.