Check for status of various vvv-phpunit dependencies before installing
Ideally we would check for version numbers as well, but the bash output just wasn't cooperating. I know this somewhat defeats the purpose of having dependencies managed by Composer itself, but it's sooooo slooooow waiting for it to hunt down updates. I'd rather tell it to only update when a version number does not yet exist.