Set WP_DEBUG with wp-cli extra-php parameter
Now that we're updating wp-cli on provision, we have access to the new extra-php parameter in wp-cli. This allows us to add PHP code to the wp-config.php file as it is created by wp-cli through the use of here-docs.

Here-docs are ugly and screw with the indentation of the provision.sh file, but that's cool. It might be worth piping in PHP from a file if possible and if we ever get to the point where we need something more complicated set.

Closes #81