Make server private key generation quieter
With the way that Vagrant works when scrolling screens through, the private key generation ends up being a looooooong series of red dots. By storing the command as we issue it, we can provide a nice one or two line output as we go instead.

We'll apply this to other commands as well, this just happens to be the current longest offender.