Add secondary IP to ping for network connectivity test
In the chance that Google's primary DNS server does not reply to a ping properly, also check one of Level3's DNS servers. If both are not replying, then we can assume (more accurately than before) that a full Internet connection is not available.

Related to #107, though ideas are still welcome.