Handle mysql-server in the same flow as other packages
We had a separate check for mysql-server under the thought that we would only set the debian config for the root password when needed. This happens so quickly that it doesn't really matter if it runs during every provision. Accepting that allows us to move the mysql-server package into the previous logic to cut down on a bit of time and help with organization.