Check IP of VM after ack-grep is installed
On a clean `vagrant up` and error was being generated because ack-grep was not yet installed when we store the VM's IP address. Moving this check further down the provisioning script until we either install ack-grep earlier or use something else to parse it. Also, we only use the variable once, so no big deal right now. :)