Remove poor implementation of Watchr
Having a watchr script was great in concept, but not entirely necessary. Now that we've improved our provisioning time for existing setups, typing `vagrant provision` at the command line initiates a process that only takes 11 or 12 seconds. Automated restarting of services on config file changes is no longer really necessary.

If we do introduce this or something like it in the future, we need to make sure it falls in the 'very useful' category.

Closes #44