Introduce support for the Vagrant hostsupdater plugin
If the hostsupdater plugin is installed, we add the domains used by VVV to the local machine's hosts file. If the hostsupdater plugin is not installed, no hosts file changes are made and no errors are thrown.

In the near term we'll want to work on a way to pull these aliases in from some kind of config file that can be maintained locally so that Vagrantfile can be updated with ease.

For now - if you'd like to add more aliases and don't want to worry about merging new commits from the master repo, the best bet may be to maintain a Customfile with an additional config line for `config.hostsupdater.aliases`.

Addresses #140, may likely fix it. I was too lazy to make a pull request, I'm sorry. :)