Remove our custom Webgrind config.php file in favor of the default
The Webgrind repository provides a full config.php file rather than a sample that is not immediately used. For this reason, whenever we symlink or copy our custom config.php file over that one in our git clone of the repo, git thinks that we have locally modified changes and a small notice pops out during provisioning. See #118

Separately... Virtualbox, at least on some OSes doesn't do symlinking between two shared folders inside the VM. This is currently causing issues in Windows, and quite possibly could appear in other places as well. See #120 and https://www.virtualbox.org/ticket/10085

Back to the first issue, it's entirely possible that an updated config.php file from the repository could overwrite our modifications **OR** that we could all of a sudden have incompatible modifications. Mostly for this reason, but also to clean up the others, and because the only change we were making to config.php was the timezone, I'm removing the custom file in favor of the repository default.

Closes #118, #120