Update Vagrant's SSH timeout time to 80 seconds
By default this value is 30 seconds, and most often we probably don't need to rely on it at all. It doesn't seem bad to have a higher value on a local setup though, just in case it does come in handy.

See #136 for how this was a possible fix earlier.