Add note on future support of nginx-permalink-fix.php
Due to http://core.trac.wordpress.org/ticket/25098, we'll no longer need this Nginx fix file hanging around in the VVV repo. It should stay for now because the default WordPress install will continue to be 3.6. Once 3.7 is released, then we can remove this from the repo and celebrate the nice things that we have been given. :)

See #156