Update provision.sh to handle default and custom vvv-hosts.data files.
Was having trouble getting the existing stuff in provision.sh to process all the vvv-hosts.dat files, so I went with a while loop instead, reading the output of the find command as if it was a file. This now checks to see if an exact match already exists before adding it to the hosts file, lest we end up with a huge file after a bunch of provisions.

I don't understand all of the magic, but there's some fanciness in line 507 that allows it to read each line and ignore empty lines.

Also clarifies a few things in the comments and adds some output to the provisioning explaining what domains were added and from what hosts files.