Remove unnecessary piping to echo statements
There are a bunch of places where we pipe a `ln` or `cp` to an `echo` with a message about what just happened. In a few places this causes readability issues as the command is so long that it wraps to the next line.

This commit ditches that and puts the `echo` on the next line down to make things more readable.