Modify the README a bit to clarify some steps and goals
Biggest change is in the VVV goals, specifically around automatic site creation, which we're working toward in #100. I don't think VVV is the right place near term for determining the best way to automatically describe WordPress projects as these are so tied to various production systems and VVV has become less production oriented.

Easing the process of creating sites on the fly will give us clarity for the future.