Clarify Vagrant, Virtualbox, and vagrant-hostsupdater versioning and usage
We live in a world of rapidly changing DevOps tools and both Vagrant and Virtualbox can cause their share of issues when we immediately jump to the latest release. However, jumping to the latest release also helps us be aware of issues. These changes are an attempt to clarify some of the version requirements shown in VVV, which aren't always requirements but helpful pointers to current versions. Closes #190

Because this is an overall README change, I also snuck in some additional password instructions around vagrant-hostsupdater. Closes #194