Copy config files instead of linking
In order to start services on init and have them work as expected, we need our custom config files to be available. Provisioning should be used to put these config files into a certain state on the development machine, not necessarily to maintain a constantly linked version. By copying the files instead of linking, we are able to use the config files before the drives are mapped later in the provisioning process and the services are able to start properly.