Copy individual nginx site configs rather than linking
Still some work to do on this as existing setups may completely blow up when we remove linked files before then copying the non-existent files. It'd be nice to find a way to unlink things first. Hrm. We're probably going to need to recommend a vagrant destroy for this update. We'll see.