Unlink previously symlinked files before copying
This sets things up to work on future uses of halt/up without requiring a full destroy. As long as `vagrant provision` is run, no destroy is needed.

Note that it doesn't make sense to keep this here forever. At some point we'll just have to recommend a full `vagrant destroy` to sync back up with the repo.