Implement style guide recommendations for if/for/while statements
From the guide:

"Put ; do and ; then on the same line as the while, for or if.

Loops in shell are a bit different, but we follow the same principles as with braces when declaring functions. That is: ; then and ; do should be on the same line as the if/for/while. else should be on its own line and closing statements should be on their own line vertically aligned with the opening statement."

Thinking of `; then` or `; do` as an opening bracket–`{`–for the statement makes a lot of sense when trying to grasp this in the mind of a day to day PHP dev.