Test for an empty string with `-z` rather than `-e`
Fixes the adding of vvv-hosts entries to the virtual machine, which has been broken since b669fd29e4 during an attempt to get better at empty and not empty checking.

In looking at this issue, it appears that `-e` is in use in a few places and I'm not entirely sure that it actually does anything. I'll fix those as well.