Remove defaults from my.cnf
Lots of defaults were set in the my.cnf file and as mysql versions increase, the syntax for the settings changes.  As we're  not purposely setting these values to anything particular, I've just deleted them in the my.cnf, which will use whatever default the current version of MySQL has.  I've left a few variables set in here that are relevant, such as max_allowed_packet.  I decided to remove things rather than comment them out because if you need that setting in this file, you know what you are doing enough to know where to put it without having a guide.  Having the values in there does more of a disservice I believe as it provides suggestions that have no relation to what you are trying to accomplish.  The items that are commented out and left in are things that could be toggled on and off to aide development. 

This is untested, but I'd be surprised if it caused any problems.