Remove all unlink statements in provisioning
In v1.0 of VVV, we made the change to copy a bunch of config files over rather than using symlinks. Because of this we had to unlink all of the previous symlinks before the copies would actually work. The unlinking should be treated as a temporary solution as it is somewhat confusing to run across in the provisioning script.

This commit removes unlinking. If for some reason an upgrade is done from v0.9 directrly to v1.1 or master, it is possible that some errors will appear explaining that a file cannot be copied on top of itself. This isn't much of a big deal, but should be treated with a `vagrant destroy` and `vagrant up` to clear things a bit.