Process subdomains for each default domain we provide Nginx configs for.
This makes it easier to setup multisite with subdomains while avoiding too many file edits in the VVV base. Other modifications will still be required, such as setting up recognition for the new hosts on the local machine and in the guest, but this takes away the need to restart Nginx.