Move back to the ondrej/php55 PPA for PHP 5.5
The move to PHP 5.5 as stable in ondrej surprised us a long while
ago, but not we (and WordPress core) are better prepared.

This will cause the latest stable of PHP 5.5 to install during
provisioning if the package is not yet installed. To force,
a vagrant destroy and vagrant up should be used.