Remove php-apc package installation
PHP 5.5 brings us a built in Zend Opcache and it does not appear
that APC has been deemed compatible with PHP 5.5. Either way, it's
on the way out and we should drop it.

This means that anything relying on APC for an object cache will
not work. See also -> we install memcache.