Enable opcache and set initial memory limit to 128
This mimics the settings that were in place for APC.

We can address the log settings with another log ticket, as it
would be interesting to have a specific log file for opcache
errors, especially because it is new.

Some of the other settings are interesting enough to leave in this
file, but should probably only be changed on a project by project
basis until we know more.