Add triggers for destroy and suspend
We can add triggers for a bunch of different Vagrant commands, but
the ones that bring down a VM seem most important in the near term
for things like backing up databases. This commit adds support for
both `vagrant destroy` and `vagrant suspend`. We can revisit
`vagrant up`, `vagrant resume`, etc... in the future.