Introduce db_backup command to backup databases
Introduces a home/vagrant/bin/db_backup that creates SQL dumps for
each of the databases setup in MySQL. These files overwrite existing
previous versions in the database/backups directory in the local
environment and can be used to restore the database if a VM is lost
or purposely destroyed.

By default, db_backup is fired whenever the halt, suspend, or destroy
commands are used. It can also be used from the command line inside
the virtual machine or on the host machine through something like:

`vagrant ssh -c "db_backup"`