Use HTTP for network detection rather than ping
Relying on ping to detect network connectivity was troublesome as
some networks block icmp and would generate a false error. As we
need HTTP access throughout provisioning anyway, we can rely on
it as an indicator of network health.

This ends up being quite a bit faster too. We could probably get
even better with just a DNS request, but that may hide another
issue depending on the network.

Fixes #428