Use single files per table
When adding databases to the engine, MySQL adds them all to
the _same_ file by default (in < 5.6.6). When removing these databases
later, MySQL fails to truncate the file, leading to used disk space that's
not actually being used. By specifying one file per table, end users can
more easily free disk space when projects are archived/removed.

Also note, this is a feature that is on by default in MySQL >= 5.6.6, so
we might as well bake it in now to get the management benefits and get
everyone ready for an eventual version update.