Add the vagrant user to the www-data group for xdebug log access
When running php (or phpunit) inside the VM rather than the web
interface, an xdebug-remote.log access error would appear in some
scenarios. In these cases, php runs as the vagrant user rather than
the www-data user and does not have access to the log files.

This modifies the vagrant user during provisioning so that one of
its alternate groups is www-data.

See #691, #631.
Fixes #621.