Allow opcache to be easily enabled/disabled
As it stands, OPCache doesn't even work since the module isn't loaded. Here we load the extension, and move the opcache.ini to the `mods-available` folder. This allows it to be enabled/disabled from the command-line, without a having to reprovision. I've also added opcache_(on/off) scripts, but these may not be wanted, as OPCache will likely be switched on/off less than XDebug.