Add NodeSource key before package install
`nodejs` is a package we install by default. Adding the NodeSource key early (if is doesn't exist), outside of the `apt_package_install_list` check, ensures that `apt-get udpate` won't complain. It also means that users won't have to run any other commands to get the latest NodeJS outside of the `apt-get` commands.