Switch from Hostsupdater to Ghost for hostfiles
Use the vagrant-ghost plugin to consolidate and manage hostfiles
instead of the current vagrant-hostsupdater. This doesn't take
advantage of the hosts collection that Ghost provides, since it only
looks for the `aliases` file for additional hostnames. Instead, it
keeps the existing hosts array generation, then passes that array
to the Ghost config for further management.

See #606.