Remove WordPress trunk from default provisioning
This has been around since the beginning of VVV, when the develop
repos had not yet been created. There's not much sense in providing
both the legacy trunk repo and the modern develop repo.

We'll still want to document the best way to continue using the
wordpress-trunk directory if it exists in your VVV instance.

See #921.