Checkout and initialize wordpress-develop in /tmp/
The installation of local npm packages was running into trouble
due to what is likely a race condition caused by slow shared drives
between the host and guest machines. This is mostly a guess, but
seems plausible.

We can instead handle the initial SVN checkout, the npm package
installation, and the initial build via grunt in the VM's `/tmp`
machine. Once this is all prepped, it is then moved to the final
location in `/srv/www/wordpress-develop` before other scripts that
install WordPress run.

Anything involving files runs quite a bit quicker during the initial
setup. The move of the directory takes a long while though, which
kind of negates the original speed. :)

But speed isn't the problem, incomplete local npm packages is.

Fixes #825.