Add the private status to available default status.
When using WP Idea Stream with a wp-idea-stream-custom.php file, we were using a filter to force the talk status to be private. The private status is interesting because Admins can read all private posts and regular users only their posts. To avoid using a custom "speaker" role, and add more flexibility so that it is still possible to organise a public call for speakers i will use the default talk status to map capabilities according to its value. Eg: private status will act like what we used to do for WordCamp Paris.

See #1"