Add the workflow state feature.
This allows to manage a mini workflow about talks. By default all talks are pending. From the Administration screens, it is possible to edit the workflow state to filter the list of talks according to workflow state. Eg: filter the list by "selected" state to get all selected talks.

See #1