Group hierarchical categories in the submit form.
NB: this only support a 1 level hierarchy.

Categories having the same parent will be grouped into the submit form according to this parent. The parent category becomes the checkbox group title. If the parent category has a description it is used to output a complementary information about the checkbox group.
If there are "flat" categrories they will also be output, just like before.

Fixes #5