Blind raters are now completely blind :)
Introduce 2 new capabilities to check the user can view rates and comments and the corresponding mapping in case the call for speakers is completely public.
Admins and regular raters can still view any comment and any rating.
Blind raters can only view their comments and their rates. I also removed the order choices as it was a way to have an idea about the rates.

Fixes #8