Adapt comment reply link & edit post link
If the user who commented the talk is a blind rater, the reply link is to replace as the blind rater will not be able to see the reply. In case the current user is a blind rater, simply remove the reply link. Otherwise display a message to inform the user will not be able to view the reply.
Adapt the capability check of the edit post link so that Jurors can edit the talk within the edit.php administration screen.