Use direct package names rather than alias names
The command we use to check if a package is installed, `dpkg -s`,
does not appear to properly follow alias names. After initial
provision, `dpkg` fails to see that several packages are already
installed and then falsely flags them to be installed.

By shifting to the original package name, provisioning should
be a bit quieter when reused.