Update README
A lot of documentation is incorrect and extremely confusing for fresh installers/new users of VM. 
If fresh install downloads VB 6.1.x [Windows], VVV sets default provider to Hyper-V instead of VB, even if the Windows OS is compatible with Hyper-V.

The only way of being aware that VB 6.1.x is incompatible is by adding --provider=virtualbox at the end of vagrant up, to which we will receive the error explaining that it only is compatible with 6.0 and below. There is no documentation of this.